+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## [Release Alpha 1.4](https://github.com/Nevah5/wordlebot/releases/tag/a1.4) - 08.02.2022
 ### Added
 - Mysql module and integration into db module
+- Query to sql db to get current game of user
 
 ### Changed
 - Example environment file, replaced port with database name